// contracts/generated/typescript/models/PageCollection.ts
export interface PageCollection {
    /**
     *
     * @type {string}
     * @memberof PageCollection
     */
    id: string;
    /**
     *
     * @type {string}
     * @memberof PageCollection
     */
    wikiId: string;
    /**
     *
     * @type {PageCollectionCollectionTypeEnum}
     * @memberof PageCollection
     */
    collectionType: PageCollectionCollectionTypeEnum;
    /**
     *
     * @type {string}
     * @memberof PageCollection
     */
    title: string;
    /**
     *
     * @type {string}
     * @memberof PageCollection
     */
    descriptionPageId: string;
    /**
     *
     * @type {CollectionQuery}
     * @memberof PageCollection
     */
    query: CollectionQuery;
    /**
     *
     * @type {Date}
     * @memberof PageCollection
     */
    createdAt: Date;
    /**
     *
     * @type {Date}
     * @memberof PageCollection
     */
    updatedAt: Date;
    /**
     * 页面命中该 Collection 的路径
     * @type {PageCollectionMembershipSourceEnum}
     * @memberof PageCollection
     */
    membershipSource: PageCollectionMembershipSourceEnum;
}


/**
 * @export
 */
export const PageCollectionCollectionTypeEnum = {
    Manual: 'manual',
    Rule: 'rule',
    Dynamic: 'dynamic'
} as const;
export type PageCollectionCollectionTypeEnum = typeof PageCollectionCollectionTypeEnum[keyof typeof PageCollectionCollectionTypeEnum];

/**
 * @export
 */
export const PageCollectionMembershipSourceEnum = {
    Page: 'page',
    PrimaryEntity: 'primary_entity',
    DynamicPage: 'dynamic_page',
    DynamicEntity: 'dynamic_entity'
} as const;
export type PageCollectionMembershipSourceEnum = typeof PageCollectionMembershipSourceEnum[keyof typeof PageCollectionMembershipSourceEnum];